+++ src/main/java/no/gunbang/market/common/config/RedisConfig.java
@@ -1,10 +1,9 @@
-package no.gunbang.market.common;
+package no.gunbang.market.common.config;
 
 import io.lettuce.core.RedisClient;
 import org.redisson.Redisson;
 import org.redisson.api.RedissonClient;
 import org.redisson.config.Config;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.session.data.redis.config.annotation.web.http.EnableRedisHttpSession;
